requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
cloudinary==1.41.0
crispy-bootstrap4==2024.10
dj-database-url==2.3.0
Django==5.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-crispy-forms==2.3
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
psycopg2==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
python-decouple==3.8
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
whitenoise==6.8.2
